Make generated salaries follow gaussian distribution

+++ src/main/java/com/anzisolutions/bankingsimulator/client/Finances.java
@@ -26,7 +26,7 @@ public class Finances {
 		return cash;
 	}
 
-	public void payday(int salary) {
+	public void payday(long salary) {
 		cash += salary;
 		incomeTotal += salary;
 	}
